dev_requirements.txt
-r requirements.txt
sphinx
sphinx-argparse
sphinx_rtd_theme
coverage<5.0
flake8>=3.7
numpydoc
pytest
pytest-cov
requests
sphinx
